% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfu_aggregations.R
\name{calculate_pfu_aggregations}
\alias{calculate_pfu_aggregations}
\title{Create a tidy data frame containing aggregate PFU data}
\usage{
calculate_pfu_aggregations(
  .sutdata,
  fd_sectors = IEATools::fd_sectors,
  p_industry_prefixes = get_p_industry_prefixes(),
  e_dot = IEATools::iea_cols$e_dot,
  sector_colname = PFUAggDatabase::sea_cols$sector_colname,
  flow_colname = PFUAggDatabase::sea_cols$flow_colname,
  agg_by_colname = PFUAggDatabase::sea_cols$agg_by_colname,
  ex_colname = PFUAggDatabase::sea_cols$ex_colname
)
}
\arguments{
\item{.sutdata}{A PSUT data frame}

\item{fd_sectors}{A list of final demand sectors, usually equal to \code{IEATools::fd_sectors}.}

\item{p_industry_prefixes}{A list of primary industry prefixes, usually produced by
calling the \code{get_p_industry_prefixes()} function.}

\item{sector_colname, flow_colname, agg_by_colname, ex_colname}{See \code{PFUAggDatabase::sea_cols}.}

\item{e_dot_colname}{See \code{IEATools::iea_cols}.}
}
\description{
Create a tidy data frame containing aggregate primary, final, and useful
energy and exergy data grouped by total, product, and sector.
}
\examples{
library(Recca)
agg_pfu <- Recca::UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = matrix.name,
                     values_from = matrix) \%>\%
  dplyr::mutate(Method = "PCM") \%>\%
  calculate_pfu_aggregations()
}
