% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{create_agg_graph}
\alias{create_agg_graph}
\title{Generate an mw graph}
\usage{
create_agg_graph(
  .df,
  country_val,
  method_val,
  energy_type_val,
  gross_net_val,
  country_colname = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  e_dot = IEATools::iea_cols$e_dot,
  stage_colname = PFUAggDatabase::sea_cols$stage_colname,
  agg_by_colname = PFUAggDatabase::sea_cols$agg_by_colname,
  sector_colname = PFUAggDatabase::sea_cols$sector_colname,
  e_product_colname = PFUAggDatabase::sea_cols$e_product_colname,
  method_colname = IEATools::iea_cols$method,
  energy_type_colname = IEATools::iea_cols$energy_type,
  gross_net_colname = PFUAggDatabase::sea_cols$gross_net_colname,
  ...
)
}
\arguments{
\item{.df}{A data frame comprised of aggregated muscle work data, usually
retrieved from the \code{...} target.}

\item{country_val}{The country for which plots are to be created.}

\item{country, year, e_dot}{See \code{IEATools::iea_cols}.}
}
\value{
A \code{ggplot2} graph object
}
\description{
Creates an mw graph from a MW PSUT.
}
\details{
This function is called repeatedly from \code{mw_plots_df()}.
}
